--- src/main/java/school/faang/user_service/controller/UserController.java
@@ -1,21 +1,0 @@
-package school.faang.user_service.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import school.faang.user_service.service.UserService;
-
-@RequiredArgsConstructor
-@RequestMapping("/user")
-@RestController
-public class UserController {
-    private final UserService userService;
-
-    @DeleteMapping("/deactivate")
-    public void deactivateUser(@RequestParam("userId") Long userId) {
-        userService.deactivateUser(userId);
-    }
-}
