+++ src/main/java/school/faang/user_service/service/user/UserErrorMessage.java
@@ -1,0 +1,6 @@
+package school.faang.user_service.service.user;
+
+public final class UserErrorMessage {
+    public static final String USER_NOT_FOUND = "User with this id not found: %s";
+    public static final String USERS_NOT_FOUND = "Users with this ids not found";
+}
